Rename google-suite provider to google-worspace
