+++ providers/google/src/airflow/providers/google/cloud/transfers/gdrive_to_gcs.py
@@ -21,7 +21,7 @@ from typing import TYPE_CHECKING
 
 from airflow.models import BaseOperator
 from airflow.providers.google.cloud.hooks.gcs import GCSHook
-from airflow.providers.google.suite.hooks.drive import GoogleDriveHook
+from airflow.providers.google.workspace.hooks.drive import GoogleDriveHook
 
 if TYPE_CHECKING:
     from airflow.utils.context import Context
