+++ providers/google/src/airflow/providers/google/workspace/sensors/drive.py
@@ -1,3 +1,4 @@
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,26 +15,29 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module contains Google Drive sensors."""
+
 from __future__ import annotations
 
 from collections.abc import Sequence
-from typing import Any
+from typing import TYPE_CHECKING
 
-from airflow.models import BaseOperator
-from airflow.providers.google.suite.hooks.sheets import GSheetsHook
+from airflow.providers.google.workspace.hooks.drive import GoogleDriveHook
+from airflow.sensors.base import BaseSensorOperator
+
+if TYPE_CHECKING:
+    from airflow.utils.context import Context
 
 
-class GoogleSheetsCreateSpreadsheetOperator(BaseOperator):
+class GoogleDriveFileExistenceSensor(BaseSensorOperator):
     """
-    Creates a new spreadsheet.
+    Checks for the existence of a file in Google Cloud Storage.
 
-    .. seealso::
-        For more information on how to use this operator, take a look at the guide:
-        :ref:`howto/operator:GoogleSheetsCreateSpreadsheetOperator`
-
-    :param spreadsheet: an instance of Spreadsheet
-        https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets#Spreadsheet
-    :param gcp_conn_id: The connection ID to use when fetching connection info.
+    :param folder_id: The Google drive folder where the file is.
+    :param file_name: The name of the file to check in Google Drive
+    :param drive_id: Optional. The id of the shared Google Drive in which the file resides.
+    :param gcp_conn_id: The connection ID to use when
+        connecting to Google Cloud Storage.
     :param impersonation_chain: Optional service account to impersonate using short-term
         credentials, or chained list of accounts required to get the access_token
         of the last account in the list, which will be impersonated in the request.
@@ -45,29 +49,34 @@ class GoogleSheetsCreateSpreadsheetOperator(BaseOperator):
     """
 
     template_fields: Sequence[str] = (
-        "spreadsheet",
+        "folder_id",
+        "file_name",
+        "drive_id",
         "impersonation_chain",
     )
+    ui_color = "#f0eee4"
 
     def __init__(
         self,
         *,
-        spreadsheet: dict[str, Any],
+        folder_id: str,
+        file_name: str,
+        drive_id: str | None = None,
         gcp_conn_id: str = "google_cloud_default",
         impersonation_chain: str | Sequence[str] | None = None,
         **kwargs,
     ) -> None:
         super().__init__(**kwargs)
+        self.folder_id = folder_id
+        self.file_name = file_name
+        self.drive_id = drive_id
         self.gcp_conn_id = gcp_conn_id
-        self.spreadsheet = spreadsheet
         self.impersonation_chain = impersonation_chain
 
-    def execute(self, context: Any) -> dict[str, Any]:
-        hook = GSheetsHook(
+    def poke(self, context: Context) -> bool:
+        self.log.info("Sensor is checking for the file %s in the folder %s", self.file_name, self.folder_id)
+        hook = GoogleDriveHook(
             gcp_conn_id=self.gcp_conn_id,
             impersonation_chain=self.impersonation_chain,
         )
-        spreadsheet = hook.create_spreadsheet(spreadsheet=self.spreadsheet)
-        self.xcom_push(context, "spreadsheet_id", spreadsheet["spreadsheetId"])
-        self.xcom_push(context, "spreadsheet_url", spreadsheet["spreadsheetUrl"])
-        return spreadsheet
+        return hook.exists(folder_id=self.folder_id, file_name=self.file_name, drive_id=self.drive_id)
