+++ providers/google/src/airflow/providers/google/workspace/transfers/gcs_to_gdrive.py
@@ -26,7 +26,7 @@ from typing import TYPE_CHECKING
 from airflow.exceptions import AirflowException
 from airflow.models import BaseOperator
 from airflow.providers.google.cloud.hooks.gcs import GCSHook
-from airflow.providers.google.suite.hooks.drive import GoogleDriveHook
+from airflow.providers.google.workspace.hooks.drive import GoogleDriveHook
 
 if TYPE_CHECKING:
     from airflow.utils.context import Context
