+++ providers/google/src/airflow/providers/google/workspace/transfers/local_to_drive.py
@@ -25,7 +25,7 @@ from typing import TYPE_CHECKING
 
 from airflow.exceptions import AirflowFailException
 from airflow.models import BaseOperator
-from airflow.providers.google.suite.hooks.drive import GoogleDriveHook
+from airflow.providers.google.workspace.hooks.drive import GoogleDriveHook
 
 if TYPE_CHECKING:
     from airflow.utils.context import Context
